Decode ARB outputs to digit vectors; dtype fixes
Convert ARB circle encodings into explicit decimal digit vectors before injection and fix dtype handling.

- ArithmeticResidualBlock: replace previous circle-output dimensionality with decimal-digit sizing (total_result_dim = num_results * num_digits + 1) and add _decode_to_digits to CRT-reconstruct operation results, detach CRT paths, convert integers to digit tensors (sub includes sign bit), and concatenate all 5 operation results into [B,S,5K+1]. The forward pass now uses these decoded digit vectors for injection and returns (h', d_a, d_b) (continuous pre-round outputs removed).
- stage4_inject: compute projection in the projection weight dtype (e.g. float32) for precision, then cast the projected delta back to the hidden state's dtype to avoid precision/gradient issues on mixed-dtype devices.
- gpt2_arb: update AutoModelForCausalLM.from_pretrained argument name from torch_dtype to dtype for HF compatibility.

These changes produce cleaner, linearly separable features for the injector, avoid gradients through CRT reconstruction, and improve numerical stability across mixed-precision hardware.

+++ mathllm/arb/stage4_inject.py
@@ -65,7 +65,10 @@ class ResultInjector(nn.Module):
         Returns:
             h': [batch, seq, hidden_dim] — h + gate * delta_h
         """
+        # Compute in the projection's dtype (float32) for gradient precision,
+        # then cast delta to match the hidden state dtype (may be float16).
+        results = results.to(dtype=self.projection.weight.dtype)
         gate = torch.sigmoid(self.gate_logit)
         delta_h = self.projection(results)
         delta_h = self.dropout(delta_h)
-        return h + gate * delta_h
+        return h + gate * delta_h.to(dtype=h.dtype)
